backend/python/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
redis==5.0.1